integrating update class on main project
